+++ admin-frontend/src/pages/AdminPage.tsx
@@ -2,23 +2,23 @@ import { Route, Routes } from "react-router-dom"
 import SideBar from "../components/SideBar"
 import styles from "./AdminPage.module.css"
 import InitPage from "./adminPages/InitPage"
-import UserManagePage from "./adminPages/UserManagePage"
-import UserDetailPage from "./adminPages/UserDetailPage"
-import BoardManagePage from "./adminPages/BoardManagePage"
-import NotifyBoardManagePage from "./adminPages/NotifyBoardManagePage"
-import NotifyBoardInsertPage from "./adminPages/NotifyBoardInsertPage"
-import InfoBoardManagePage from "./adminPages/InfoBoardManagePage"
-import InfoBoardInsertPage from "./adminPages/InfoBoardInsertPage"
-import EventBoardManagePage from "./adminPages/EventBoardManagePage"
-import EventBoardInsertPage from "./adminPages/EventBoardInsertPage"
-import ProductBoardManagePage from "./adminPages/ProductBoardManagePage"
-import ProductBoardInsertPage from "./adminPages/ProductBoardInsertPage"
-import AskManagePage from "./adminPages/AskManagePage"
-import ReportManagePage from "./adminPages/ReportManagePage"
-import NotifyBoardUpdatePage from "./adminPages/NotifyBoardUpdatePage"
-import InfoBoardUpdatePage from "./adminPages/InfoBoardUpdatePage"
-import ProductBoardUpdatePage from "./adminPages/ProductBoardUpdatePage"
-import EventBoardUpdatePage from "./adminPages/EventBoardUpdatePage"
+import UserManagePage from "./adminPages/user/UserManagePage"
+import UserDetailPage from "./adminPages/user/UserDetailPage"
+import BoardManagePage from "./adminPages/board/BoardManagePage"
+import NotifyBoardManagePage from "./adminPages/notify/NotifyBoardManagePage"
+import NotifyBoardInsertPage from "./adminPages/notify/NotifyBoardInsertPage"
+import InfoBoardManagePage from "./adminPages/info/InfoBoardManagePage"
+import InfoBoardInsertPage from "./adminPages/info/InfoBoardInsertPage"
+import EventBoardManagePage from "./adminPages/event/EventBoardManagePage"
+import EventBoardInsertPage from "./adminPages/event/EventBoardInsertPage"
+import ProductBoardManagePage from "./adminPages/product/ProductBoardManagePage"
+import ProductBoardInsertPage from "./adminPages/product/ProductBoardInsertPage"
+import AskManagePage from "./adminPages/ask/AskManagePage"
+import ReportManagePage from "./adminPages/report/ReportManagePage"
+import NotifyBoardUpdatePage from "./adminPages/notify/NotifyBoardUpdatePage"
+import InfoBoardUpdatePage from "./adminPages/info/InfoBoardUpdatePage"
+import ProductBoardUpdatePage from "./adminPages/product/ProductBoardUpdatePage"
+import EventBoardUpdatePage from "./adminPages/event/EventBoardUpdatePage"
 
 export default function AdminPage() {
     return (
