+++ admin-frontend/src/pages/adminPages/board/BoardInsertPage.tsx
