+++ admin-frontend/src/pages/adminPages/info/styles/InfoBoardManagePage.module.css
