+++ admin-frontend/src/pages/adminPages/notify/NotifyBoardInsertPage.tsx
@@ -1,16 +1,14 @@
-import axios from 'axios';
-import styles from './NotifyBoardInsertPage.module.css';
-import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../store/store';
+import styles from './styles/NotifyBoardInsertPage.module.css';
+import { initialBoard } from '../../../type/board';
+import { useNavigate } from 'react-router-dom';
+import axios from 'axios';
 
-export default function NotifyBoardUpdatePage() {
+export default function NotifyBoardInsertPage() {
 
     const navi = useNavigate();
-    const cacheBoard = useSelector((state: RootState) => state.boards);
-    const [board , setBoard] = useState(cacheBoard.oneBoard);
-
+    const [board , setBoard] = useState(initialBoard);
+    
     function handleInputChange(e:React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) {
         let {name , value} = e.target;
         setBoard({
@@ -19,13 +17,14 @@ export default function NotifyBoardUpdatePage() {
         })
     }
 
-    function notifyBoardUpdate() {
+    function notifyBoardInsert() {
+
         const updatedBoard = {
             ...board,
             boardCode: 'N'
         };
 
-        axios.post("http://localhost:8013/banju/admin/board/updateBoard" , updatedBoard)
+        axios.post("http://localhost:8013/banju/admin/board/insertNofityboard" , updatedBoard)
             .then((response) => {
                 alert(response.data.msg);
                 navi('../nonifyBoardManage');
@@ -37,13 +36,13 @@ export default function NotifyBoardUpdatePage() {
 
     return (
         <div className={styles.container}>
-            <h1 className={styles.title}>공지사항 수정</h1>
+            <h1 className={styles.title}>공지사항 작성</h1>
             <div className={styles.formContainer}>
                 <div className={styles.inputField}>
                     <label className={styles.inputLabel}>제목</label>
                     <input
                         type="text"
-                        placeholder="제목을 입력하세요"
+                        placeholder="공지사항 제목을 입력하세요"
                         className={styles.textInput}
                         onChange={handleInputChange}
                         name = "title"
@@ -53,7 +52,7 @@ export default function NotifyBoardUpdatePage() {
                 <div className={styles.inputField}>
                     <label className={styles.inputLabel}>내용 작성</label>
                     <textarea
-                        placeholder="내용을 입력하세요."
+                        placeholder="공지사항 내용을 입력하세요."
                         className={styles.textArea}
                         onChange={handleInputChange}
                         name = "content"
@@ -65,7 +64,7 @@ export default function NotifyBoardUpdatePage() {
                         navi('../nonifyBoardManage');
                         }}>취소</button>
                     <button className={styles.submitButton} onClick={() => {
-                        notifyBoardUpdate();
+                        notifyBoardInsert();
                         }}>작성</button>
                 </div>
             </div>
