+++ admin-frontend/src/pages/adminPages/notify/NotifyBoardUpdatePage.tsx
@@ -1,14 +1,16 @@
-import { useState } from 'react';
-import styles from './NotifyBoardInsertPage.module.css';
-import { initialBoard } from '../../type/board';
-import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
+import styles from './styles/NotifyBoardInsertPage.module.css';
+import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
+import { useSelector } from 'react-redux';
+import { RootState } from '../../../store/store';
 
-export default function NotifyBoardInsertPage() {
+export default function NotifyBoardUpdatePage() {
 
     const navi = useNavigate();
-    const [board , setBoard] = useState(initialBoard);
-    
+    const cacheBoard = useSelector((state: RootState) => state.boards);
+    const [board , setBoard] = useState(cacheBoard.oneBoard);
+
     function handleInputChange(e:React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) {
         let {name , value} = e.target;
         setBoard({
@@ -17,14 +19,13 @@ export default function NotifyBoardInsertPage() {
         })
     }
 
-    function notifyBoardInsert() {
-
+    function notifyBoardUpdate() {
         const updatedBoard = {
             ...board,
             boardCode: 'N'
         };
 
-        axios.post("http://localhost:8013/banju/admin/board/insertNofityboard" , updatedBoard)
+        axios.post("http://localhost:8013/banju/admin/board/updateBoard" , updatedBoard)
             .then((response) => {
                 alert(response.data.msg);
                 navi('../nonifyBoardManage');
@@ -36,13 +37,13 @@ export default function NotifyBoardInsertPage() {
 
     return (
         <div className={styles.container}>
-            <h1 className={styles.title}>공지사항 작성</h1>
+            <h1 className={styles.title}>공지사항 수정</h1>
             <div className={styles.formContainer}>
                 <div className={styles.inputField}>
                     <label className={styles.inputLabel}>제목</label>
                     <input
                         type="text"
-                        placeholder="공지사항 제목을 입력하세요"
+                        placeholder="제목을 입력하세요"
                         className={styles.textInput}
                         onChange={handleInputChange}
                         name = "title"
@@ -52,7 +53,7 @@ export default function NotifyBoardInsertPage() {
                 <div className={styles.inputField}>
                     <label className={styles.inputLabel}>내용 작성</label>
                     <textarea
-                        placeholder="공지사항 내용을 입력하세요."
+                        placeholder="내용을 입력하세요."
                         className={styles.textArea}
                         onChange={handleInputChange}
                         name = "content"
@@ -64,7 +65,7 @@ export default function NotifyBoardInsertPage() {
                         navi('../nonifyBoardManage');
                         }}>취소</button>
                     <button className={styles.submitButton} onClick={() => {
-                        notifyBoardInsert();
+                        notifyBoardUpdate();
                         }}>작성</button>
                 </div>
             </div>
