+++ admin-frontend/src/pages/adminPages/user/styles/UserDetailPage.module.css
